pyproject.toml
[project]
name = "code-review-platform"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "pylint>=3.0.0",
    "black>=24.0.0",
    "bandit>=1.7.0",
    "pytest>=8.0.0",
    "PyGithub>=2.0.0",
    "radon>=6.0.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["core", "skills", "agents", "hooks"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
